Refactors _fetch_guard() to base class
