+++ conda_recipe_manager/fetcher/base_artifact_fetcher.py
@@ -9,6 +9,8 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Final
 
+from conda_recipe_manager.fetcher.exceptions import FetchRequiredError
+
 # Identifying string used to flag temp files and directories created by this module.
 _ARTIFACT_FETCHER_FILE_ID: Final[str] = "crm_artifact_fetcher"
 
@@ -37,6 +39,18 @@ class BaseArtifactFetcher(metaclass=ABCMeta):
         self._temp_dir_path: Final[Path] = Path(self._temp_dir.name)
         # Flag to track if `fetch()` has been called successfully once.
         self._successfully_fetched = False
+
+    def _fetch_guard(self, msg: str) -> None:
+        """
+        Convenience function that prevents executing functions that require the code to be downloaded or stored to the
+        temporary directory.
+
+        :param msg: Message to attach to the exception.
+        :raises FetchRequiredError: If `fetch()` has not been successfully invoked.
+        """
+        if self._successfully_fetched:
+            return
+        raise FetchRequiredError(msg)
 
     @abstractmethod
     def fetch(self) -> None:
